Fix opkg-build to handle errors correctly
We need opkg-build to fail if for example the tar command is passed invalid
options. Without this, we see silently created empty packaged where data.tar
is zero bytes in size. This creates hard to debug problems.

An example is when reproducible builds are enabled and run on old hosts like
centos7 which has tar < 1.28:

Subprocess output:tar: unrecognized option '--clamp-mtime'
Try `tar --help' or `tar --usage' for more information.

This change does make the script bash specific but that shouldn't an issue
for systems capable of building ipks.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>