opkg-make-index: make use of multithreading
The call to pkg.print can be parallelised to save time as this
call is expensive since it triggers md5/sha256 operations. This
commit adds a multiprocessing pool to parallelise the pkg.print
operations and adds a post-processing step to print the outputs.

This commit adds the option -t to the command parameters allowing
a user to specify the number of threads to use. Defaults to 1 which
turns off any multithreading activity.

A quick benchmark using roughly 2000 ipks (with 2M of contents)
and on a dual core laptop showed a speed improvement of 50%, from
22 seconds to 11 seconds (with no checksums enabled) and 47 seconds
to 22 seconds (with md5 and sha256 checksums enabled).

Signed-off-by: Michael Ho <michael.ho@ieee.org>